requirements.txt
selenium==4.15.1
pytest==7.0.1
allure-pytest==2.9.45
allure-python-commons~=2.9.45
pytest-html==3.1.1
webdriver-manager==3.8.5
requests==2.32.0
undetected-chromedriver~=3.5.5
Faker~=37.1.0